Refactor: Shielded hardsuits (#1194)
* feat: Рефактор щитов-хардсьютов

* fix: Отображение щит-хардсьют мага при зарядке

* fix: Правильный путь до коробки

* fix: Правильный путь до шлема

+++ code/modules/clothing/suits/wiz_robe.dm
@@ -174,9 +174,7 @@
 	resistance_flags = FLAMMABLE
 	magical = FALSE
 
-//Shielded Armour
-
-/obj/item/clothing/suit/space/hardsuit/shielded/wizard
+/obj/item/clothing/suit/space/hardsuit/wizard
 	name = "battlemage armour"
 	desc = "Not all wizards are afraid of getting up close and personal. Not spaceproof despite its appearance."
 	icon_state = "hardsuit-wiz"
@@ -185,26 +183,12 @@
 	min_cold_protection_temperature = ARMOR_MIN_TEMP_PROTECT
 	max_heat_protection_temperature = ARMOR_MAX_TEMP_PROTECT
 	resistance_flags = FIRE_PROOF | ACID_PROOF
-	recharge_rate = 0
-	current_charges = 15
-	recharge_cooldown = INFINITY
-	shield_state = "shield-red"
-	shield_on = "shield-red"
-	helmettype = /obj/item/clothing/head/helmet/space/hardsuit/shielded/wizard
+	helmettype = /obj/item/clothing/head/helmet/space/hardsuit/wizard
 	flags_inv = HIDESHOES|HIDEJUMPSUIT|HIDETAIL
 	slowdown = 0
 	magical = TRUE
 
-/obj/item/clothing/suit/space/hardsuit/shielded/wizard/arch
-	desc = "For the arch wizard in need of additional protection."
-	recharge_rate = 1
-	recharge_cooldown = 0
-	max_charges = 15
-	min_cold_protection_temperature = SPACE_SUIT_MIN_TEMP_PROTECT
-	max_heat_protection_temperature = SPACE_SUIT_MAX_TEMP_PROTECT
-	helmettype = /obj/item/clothing/head/helmet/space/hardsuit/shielded/wizard/arch
-
-/obj/item/clothing/head/helmet/space/hardsuit/shielded/wizard
+/obj/item/clothing/head/helmet/space/hardsuit/wizard
 	name = "battlemage helmet"
 	desc = "A suitably impressive helmet."
 	icon_state = "hardsuit0-wiz"
@@ -213,14 +197,33 @@
 	item_color = "wiz"
 	min_cold_protection_temperature = ARMOR_MIN_TEMP_PROTECT
 	max_heat_protection_temperature = ARMOR_MAX_TEMP_PROTECT
-	resistance_flags = FIRE_PROOF | ACID_PROOF	
+	resistance_flags = FIRE_PROOF | ACID_PROOF
 	actions_types = list() //No inbuilt light
 	magical = TRUE
 
-/obj/item/clothing/head/helmet/space/hardsuit/shielded/wizard/attack_self(mob/user)
+//Shielded Armour
+
+/obj/item/clothing/suit/space/hardsuit/wizard/shielded
+	shielded = TRUE
+	recharge_rate = 0
+	current_charges = 15
+	recharge_cooldown = INFINITY
+	shield_state = "shield-red"
+	shield_on = "shield-red"
+
+/obj/item/clothing/head/helmet/space/hardsuit/wizard/shielded/attack_self(mob/user)
 	return
 
-/obj/item/clothing/head/helmet/space/hardsuit/shielded/wizard/arch
+/obj/item/clothing/suit/space/hardsuit/wizard/shielded/arch
+	desc = "For the arch wizard in need of additional protection."
+	recharge_rate = 1
+	recharge_cooldown = 0
+	max_charges = 15
+	min_cold_protection_temperature = SPACE_SUIT_MIN_TEMP_PROTECT
+	max_heat_protection_temperature = SPACE_SUIT_MAX_TEMP_PROTECT
+	helmettype = /obj/item/clothing/head/helmet/space/hardsuit/wizard/shielded/arch
+
+/obj/item/clothing/head/helmet/space/hardsuit/wizard/shielded/arch
 	desc = "A truly protective helmet."
 	min_cold_protection_temperature = SPACE_HELM_MIN_TEMP_PROTECT
 	max_heat_protection_temperature = SPACE_HELM_MAX_TEMP_PROTECT
@@ -231,11 +234,16 @@
 	icon = 'icons/effects/effects.dmi'
 	icon_state = "electricity2"
 
-/obj/item/wizard_armour_charge/afterattack(obj/item/clothing/suit/space/hardsuit/shielded/wizard/W, mob/user)
+/obj/item/wizard_armour_charge/afterattack(obj/item/clothing/suit/space/hardsuit/wizard/shielded/W, mob/user)
 	. = ..()
 	if(!istype(W))
 		to_chat(user, "<span class='warning'>The rune can only be used on battlemage armour!</span>")
 		return
+	if(W == user.get_item_by_slot(slot_wear_suit))
+		to_chat(user, "<span class='warning'>You cannot replenish charges to [W] while wearing it.</span>")
+		return
 	W.current_charges += 8
+	W.shield_state = "[W.shield_on]"
+	playsound(loc, 'sound/magic/charge.ogg', 50, TRUE)
 	to_chat(user, "<span class='notice'>You charge [W]. It can now absorb [W.current_charges] hits.</span>")
 	qdel(src)
